{"types":"C","runs":49}

+++ Practice.java
@@ -5,7 +5,7 @@ import acm.program.ConsoleProgram;
 public class Practice extends ConsoleProgram {
 	public void run(){
 		
-		double n = readDouble();
+		int n = readInt();
 		for(int i = 0; i < 0; i++){
 			println("hello" + i);
 		}
